Add support for openapi schema generation
Add worf.views.schema view that can be used to auto-generate the openapi schema for the url patterns defined in settings.ROOT_URLCONF. The view can be extended to generate the schema with different title/base_url, and/or custom url patterns.

The worf.views.schema  is using the newly added module worf.schema.schema_generator in order to produce a valid openapi schema, in json format. The schema_generator is using the apispec library, in order to automate the schema generation from Marshmallow schemas.

As a result, it the users of worf may know enjoy automatic openapi docs generation for their worf views.